+++ app/src/main/java/de/triology/universeadm/CasSecurityModule.java
@@ -112,6 +112,7 @@ public class CasSecurityModule extends BaseSecurityModule {
     addFilterChain("/login/cas", ANON, CAS);
     addFilterChain("/api/users", API, config(ROLES, Roles.ADMINISTRATOR));
     addFilterChain("/api/users/*", API, config(ROLES, Roles.ADMINISTRATOR));
+    addFilterChain("/api/users/import/*", API, config(ROLES, Roles.ADMINISTRATOR));
     addFilterChain("/api/groups", API, config(ROLES, Roles.ADMINISTRATOR));
     addFilterChain("/api/groups/*", API, config(ROLES, Roles.ADMINISTRATOR));
     addFilterChain("/api/account", API);