+++ app/src/main/ui/src/hooks/useSetPageTitle.ts
@@ -1,7 +1,8 @@
 import {useEffect} from "react";
+import {t} from "../helpers/i18nHelpers";
 
 export function useSetPageTitle(title: string) {
     useEffect(() => {
-        (document.title = title);
-    }, []);
+        (document.title = `${title} | ${t("general.applicationName")}`);
+    }, [title]);
 }