+++ app/src/main/ui/src/services/CasUser.ts
@@ -3,6 +3,7 @@ import {Axios} from "../api/axios";
 export type CasUser = {
     principal: string;
     admin: boolean;
+    loading: boolean;
 }
 
 export const CasUserService = {