+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Pages only available to users with manager group will now show an error message if a normal user accesses the page (#105)
+
 ### Fixed
 - Import multiple user (>100) via csv (#101)
+- Users without manager group can access restricted areas (#105)
+- Pagination for users, groups and import summaries
 
 ## [v1.11.0-1] - 2023-09-18
 ### Added